fix: MambaBlock 残差加法改在 float32 做，防止训练后期 NaN
根因：训练几轮后 SSM 输出幅值增大，out_proj(y) 在 float32 下
超过 float16 上限（65504），.to(float16) 变 inf，
后续 GroupNorm(inf) → NaN 扩散。

修复：out + residual 在 float32 下完成加法，再整体转回 orig_dtype，
避免单独转换时的溢出。

+++ models/mamba_translator.py
@@ -173,9 +173,10 @@ class MambaBlock(nn.Module):
         y   = selective_scan(x_, dt, A, B_ssm, C_ssm, self.D.float())
         y   = y * F.silu(z)
 
-        # 输出投影 + 残差（转回原始 dtype 保持 AMP 兼容性）
-        out = self.out_proj(y).to(orig_dtype)
-        return out + residual
+        # 输出投影 + 残差：在 float32 下相加再转回，防止训练后期
+        # out_proj 输出幅值增大超过 float16 上限（65504）→ inf → NaN
+        out = self.out_proj(y)                          # float32
+        return (out + residual.float()).to(orig_dtype)  # float32 加法再转回
 
 
 # ---------------------------------------------------------------------------
